pyproject.toml
[project]
name = "shopper"
version = "0.1.0"
description = "Personal shopping agent — give it a link, it buys the thing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.12.0",
    "click>=8.0",
    "google-genai>=1.60.0",
    "playwright>=1.58.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = ["ruff>=0.15.1", "pre-commit>=4.0"]

[project.scripts]
shop = "shopper:cli"
shop-google-smoke = "shopper_google:smoke"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["F", "I"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["shopper"]
